third_party/blink/web_tests/http/tests/devtools/sources/debugger-ui/monitor-console-command-expected.txt

Tests traceCalls(fn) console command.


Running: testSimpleMonitor
Start monitoring function.
Set timer for test function.
Console message received: function simpleTestFunction called
Stop monitoring.

Running: testSimpleMonitorWith1Arg
Start monitoring function.
Set timer for test function.
Console message received: function simpleTestFunction called with arguments: 1
Stop monitoring.

Running: testSimpleMonitorWithManyArgs
Start monitoring function.
Set timer for test function.
Console message received: function simpleTestFunction called with arguments: 1, 2, 3, 4, 5
Stop monitoring.

Running: testSimpleUnmonitor
Start monitoring function.
Set timer for test function.
Console message received: function simpleTestFunction3 called
Stop monitoring.

Running: testUnmonitorFuntionNotMonitored
